+++ CHANGELOG.md
@@ -6,13 +6,17 @@ This file is used to list changes made in each version of the aws-parallelcluste
 3.6.0
 ------
 
+**CHANGES**
+- Consider dynamic nodes with INVALID_REG flag as bootstrap failure towards the Slurm protected mode.
+  - Static nodes failing the Slurm registration are already treated as a bootstrap failure after the `node_replacement_timeout`.
+
 3.5.0
 ------
 
 **ENHANCEMENTS**
 - Add logging of compute node console output to CloudWatch from head node on compute node bootstrap failure.
 - Add validators to prevent malicious string injection while calling the subprocess module.
-  
+
 **BUG FIXES**
 - Fix an issue in clustermgtd that caused compute nodes rebooted via Slurm to be replaced if the EC2 instance status checks fail.